requirements.txt
absl-py==1.4.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.6.0
anthropic==0.21.3
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argilla==1.25.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.8.2
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
dataclasses-json==0.6.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
docker==7.0.0
docker-pycreds==0.4.0
easydict==1.11
edgar==5.4.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
faiss-cpu==1.7.4
Faker==18.9.0
fastjsonschema==2.16.3
filelock==3.11.0
FLAML==2.1.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
flatbuffers==23.3.3
fonttools==4.39.3
frozendict==2.3.8
frozenlist==1.4.1
fsspec==2024.5.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
gast==0.4.0
gensim==4.3.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.4.0
google-api-core==2.19.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
h5py==3.8.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.23.2
idna==3.4
immutabledict==4.2.0
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflection==0.5.1
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.2.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jovian==0.2.47
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
keras==2.12.0
kiwisolver==1.4.4
langchain==0.1.9
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==0.0.9
langsmith==0.1.81
libclang==16.0.0
lxml==4.9.2
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.21.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.0.4
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.7.3
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==1.35.13
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.5
outcome==1.2.0
packaging==23.2
pandas==1.5.3
pandas-stubs==2.1.4.231227
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.2.0
plotly==5.15.0
prodict==0.8.18
prompt-toolkit==3.0.38
proto-plus==1.24.0
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyautogen==0.2.11
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.14.0
pymongo==4.8.0
pyparsing==3.0.9
pyreadstat==1.2.6
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-liquid==1.12.1
pytils==0.4.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
rapidfuzz==3.0.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
safetensors==0.4.3
schedule==1.2.1
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
sec-api==1.0.16
sec-edgar-downloader==4.3.0
selenium==4.9.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==2.6.0
setproctitle==1.3.3
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
sklearn==0.0.post1
smart-open==6.3.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.31
stack-data==0.6.2
statsmodels==0.14.1
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-addons==0.23.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
threadpoolctl==3.1.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenizers==0.15.2
tools==0.1.9
torch==2.2.2
torch_geometric==2.4.0
torchtext==0.9.2
torchvision==0.17.2
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.35.2
trio==0.22.0
trio-websocket==0.10.2
typeguard==2.13.3
typer==0.9.4
types-pytz==2024.1.0.20240203
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==1.26.15
uuid==1.30
wandb==0.16.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
yarl==1.9.4
yfinance==0.2.18
zipp==3.15.0
